Boilerplating
I have added some boiler plate files that i know i will need during the process, i will add others as and when i think of them along the way. For now this is the baseline. I have looked around for a suitable API for the games, however i may have to pivot either to TV Shows instead of games or to a web scraper to get the data.